refact : RunningListActivity를 Fragment로 변경
